fix: disable /kill during desktop mode
there's no need for a kill option in desktop mode and it becomes an issue if you press the exit button on the frontend

+++ amiga-app/backend/routers/tracks.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter
 from fastapi import HTTPException
 from fastapi import Response
+from fastapi import Request
 import json
 from pathlib import Path
 
@@ -13,7 +14,11 @@ router = APIRouter()
 
 
 @router.get("/kill")
-async def kill_app():
+async def kill_app(request: Request):
+    if request.app.state.desktop:
+        raise HTTPException(
+            status_code=403, detail="Killing the app is not allowed in desktop mode."
+        )
     os.kill(os.getpid(), signal.SIGTERM)
     return Response(status_code=200, content="Server shutting down...")
